test code and agent factory SR
