+++ agential/cog/self_refine/functional.py
@@ -59,13 +59,13 @@ def _prompt_agent(
         prompt=prompt,
         additional_keys=additional_keys,
     )
-    print("<PROMPT AGENT=====================================================>")
-    print(prompt)
-    print("<PROMPT AGENT=====================================================>")
+    # print("<PROMPT AGENT=====================================================>")
+    # print(prompt)
+    # print("<PROMPT AGENT=====================================================>")
     out = llm(prompt)
-    print("<OUT AGENT=====================================================>")
-    print(repr(out.output_text))
-    print("<OUT AGENT=====================================================>")
+    # print("<OUT AGENT=====================================================>")
+    # print(repr(out.output_text))
+    # print("<OUT AGENT=====================================================>")
     return out
 
 
@@ -131,13 +131,13 @@ def _prompt_critique(
         prompt=prompt,
         additional_keys=additional_keys,
     )
-    print("<PROMPT CRITIC=====================================================>")
-    print(prompt)
-    print("<PROMPT CRITIC=====================================================>")
+    # print("<PROMPT CRITIC=====================================================>")
+    # print(prompt)
+    # print("<PROMPT CRITIC=====================================================>")
     out = llm(prompt)
-    print("<OUT CRITIC=====================================================>")
-    print(repr(out.output_text))
-    print("<OUT CRITIC=====================================================>")
+    # print("<OUT CRITIC=====================================================>")
+    # print(repr(out.output_text))
+    # print("<OUT CRITIC=====================================================>")
 
     return out
 
@@ -206,13 +206,13 @@ def _prompt_refine(
         prompt=prompt,
         additional_keys=additional_keys,
     )
-    print("<PROMPT REFINE=====================================================>")
-    print(prompt)
-    print("<PROMPT REFINE=====================================================>")
+    # print("<PROMPT REFINE=====================================================>")
+    # print(prompt)
+    # print("<PROMPT REFINE=====================================================>")
     out = llm(prompt)
-    print("<OUT REFINE=====================================================>")
-    print(repr(out.output_text))
-    print("<OUT REFINE=====================================================>")
+    # print("<OUT REFINE=====================================================>")
+    # print(repr(out.output_text))
+    # print("<OUT REFINE=====================================================>")
 
     return out
