+++ tests/cog/self_refine/strategies/test_code.py
@@ -1,7 +1,7 @@
 """Unit tests for Self-Refine code strategies."""
 
 from agential.cog.fewshots.humaneval import HUMANEVAL_FEWSHOT_EXAMPLES_POT
-from agential.cog.self_refine.output import SelfRefineStepOutput
+from agential.cog.self_refine.output import SelfRefineOutput, SelfRefineStepOutput
 from agential.cog.self_refine.prompts import (
     HUMANEVAL_CRITIQUE_FEWSHOT_EXAMPLES,
     HUMANEVAL_REFINE_FEWSHOT_EXAMPLES,
@@ -40,6 +40,7 @@ def test_generate() -> None:
     question = inst["prompt"]
     tests = f"{inst['test']}\ncheck({inst['entry_point']})"
 
+    gt_out = SelfRefineOutput(additional_info=[SelfRefineStepOutput(answer='def has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])', critique='The implementation of the `has_close_elements` function is missing the return statement. The function itself iterates through pairs of numbers in the list and checks if the absolute difference between them falls below the given threshold. However, it lacks a return statement to provide the final result of whether any two numbers are closer than the threshold.\n\nTo fix this issue, the `has_close_elements` function should include a return statement at the end to return the boolean result after checking all pairs of numbers. Here is the revised code with the return statement added:\n\n```python\ndef has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```\n\nBy adding `return` at the beginning of this line, the function will correctly return `True` if any two numbers in the list are closer than the threshold and `False` otherwise.', answer_response=Response(input_text='', output_text='```python\ndef has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```', prompt_tokens=10, completion_tokens=20, total_tokens=30, prompt_cost=1.5e-05, completion_cost=3.9999999999999996e-05, total_cost=5.4999999999999995e-05, prompt_time=0.5), critique_response=Response(input_text='', output_text='The implementation of the `has_close_elements` function is missing the return statement. The function itself iterates through pairs of numbers in the list and checks if the absolute difference between them falls below the given threshold. However, it lacks a return statement to provide the final result of whether any two numbers are closer than the threshold.\n\nTo fix this issue, the `has_close_elements` function should include a return statement at the end to return the boolean result after checking all pairs of numbers. Here is the revised code with the return statement added:\n\n```python\ndef has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```\n\nBy adding `return` at the beginning of this line, the function will correctly return `True` if any two numbers in the list are closer than the threshold and `False` otherwise.', prompt_tokens=10, completion_tokens=20, total_tokens=30, prompt_cost=1.5e-05, completion_cost=3.9999999999999996e-05, total_cost=5.4999999999999995e-05, prompt_time=0.5)), SelfRefineStepOutput(answer='from typing import List\n\ndef has_close_elements(numbers: List[float], threshold: float) -> bool:\n    """ Check if in given list of numbers, are any two numbers closer to each other than\n    given threshold.\n    >>> has_close_elements([1.0, 2.0, 3.0], 0.5)\n    False\n    >>> has_close_elements([1.0, 2.8, 3.0, 4.0, 5.0, 2.0], 0.3)\n    True\n    """\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])', critique='The implementation of the `has_close_elements` function has a logical error in the condition used to check if any two numbers in the list are closer to each other than the given threshold.\n\n1. The primary issue lies in the condition used within the `any` function:\n   ```python\n   return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n   ```\n   This condition is meant to check the absolute difference between every pair of numbers (a, b) in the list and return `True` if the difference is less than the threshold for any pair. However, the current implementation fails to consider the case when `a` and `b` are the same number (which implies a difference of 0), potentially leading to incorrect results.\n\n2. In the given test cases:\n   - The case with the numbers `[1.0, 2.0, 3.9, 4.0, 5.0, 2.2]` and threshold `0.3` is expected to return `True` since `2.2` and `3.9` are closer than the threshold.\n   - The case with the same numbers but a threshold of `0.05` is expected to return `False` since the closest numbers are not within the threshold.\n\n3. To address the issue, the condition should be modified to exclude comparing the same number with itself:\n   ```python\n   return any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n   ```\n\nBy adjusting the condition to exclude comparing a number with itself, the function will correctly identify cases where two distinct numbers in the list are closer to each other than the specified threshold, providing accurate outcomes.', answer_response=Response(input_text='', output_text='```python\nfrom typing import List\n\ndef has_close_elements(numbers: List[float], threshold: float) -> bool:\n    """ Check if in given list of numbers, are any two numbers closer to each other than\n    given threshold.\n    >>> has_close_elements([1.0, 2.0, 3.0], 0.5)\n    False\n    >>> has_close_elements([1.0, 2.8, 3.0, 4.0, 5.0, 2.0], 0.3)\n    True\n    """\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```', prompt_tokens=10, completion_tokens=20, total_tokens=30, prompt_cost=1.5e-05, completion_cost=3.9999999999999996e-05, total_cost=5.4999999999999995e-05, prompt_time=0.5), critique_response=Response(input_text='', output_text='The implementation of the `has_close_elements` function has a logical error in the condition used to check if any two numbers in the list are closer to each other than the given threshold.\n\n1. The primary issue lies in the condition used within the `any` function:\n   ```python\n   return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n   ```\n   This condition is meant to check the absolute difference between every pair of numbers (a, b) in the list and return `True` if the difference is less than the threshold for any pair. However, the current implementation fails to consider the case when `a` and `b` are the same number (which implies a difference of 0), potentially leading to incorrect results.\n\n2. In the given test cases:\n   - The case with the numbers `[1.0, 2.0, 3.9, 4.0, 5.0, 2.2]` and threshold `0.3` is expected to return `True` since `2.2` and `3.9` are closer than the threshold.\n   - The case with the same numbers but a threshold of `0.05` is expected to return `False` since the closest numbers are not within the threshold.\n\n3. To address the issue, the condition should be modified to exclude comparing the same number with itself:\n   ```python\n   return any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n   ```\n\nBy adjusting the condition to exclude comparing a number with itself, the function will correctly identify cases where two distinct numbers in the list are closer to each other than the specified threshold, providing accurate outcomes.', prompt_tokens=10, completion_tokens=20, total_tokens=30, prompt_cost=1.5e-05, completion_cost=3.9999999999999996e-05, total_cost=5.4999999999999995e-05, prompt_time=0.5)), SelfRefineStepOutput(answer='from typing import List\n\ndef has_close_elements(numbers: List[float], threshold: float) -> bool:\n    """ Check if in given list of numbers, are any two numbers closer to each other than\n    given threshold.\n    >>> has_close_elements([1.0, 2.0, 3.0], 0.5)\n    False\n    >>> has_close_elements([1.0, 2.8, 3.0, 4.0, 5.0, 2.0], 0.3)\n    True\n    """\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)', critique="The `has_close_elements` function has a logical issue in the way it checks for close elements in the given list of numbers. \n\nHere's the problem: \n\nThe function uses a nested loop with list comprehension to iterate over every pair of numbers in the list and checks if the absolute difference between them is less than the threshold. However, this implementation incorrectly compares every number with every other number in the list, including comparing a number with itself (i.e., when `i == j`). This leads to false positives when the difference between a number and itself is below the threshold, which should not be considered a pair of close elements.\n\nTo fix this issue, you should modify the condition to exclude comparisons between the same numbers by adding an additional check to ensure `i != j` before comparing the two numbers:\n```python\nreturn any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n```\n\nThis adjustment ensures that each number is only compared with distinct numbers in the list, eliminating false positives that would occur when comparing a number with itself. \n\nAfter making this change, the function should correctly check for close elements in the list according to the provided threshold.", answer_response=Response(input_text='', output_text='```python\nfrom typing import List\n\ndef has_close_elements(numbers: List[float], threshold: float) -> bool:\n    """ Check if in given list of numbers, are any two numbers closer to each other than\n    given threshold.\n    >>> has_close_elements([1.0, 2.0, 3.0], 0.5)\n    False\n    >>> has_close_elements([1.0, 2.8, 3.0, 4.0, 5.0, 2.0], 0.3)\n    True\n    """\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n```', prompt_tokens=10, completion_tokens=20, total_tokens=30, prompt_cost=1.5e-05, completion_cost=3.9999999999999996e-05, total_cost=5.4999999999999995e-05, prompt_time=0.5), critique_response=Response(input_text='', output_text="The `has_close_elements` function has a logical issue in the way it checks for close elements in the given list of numbers. \n\nHere's the problem: \n\nThe function uses a nested loop with list comprehension to iterate over every pair of numbers in the list and checks if the absolute difference between them is less than the threshold. However, this implementation incorrectly compares every number with every other number in the list, including comparing a number with itself (i.e., when `i == j`). This leads to false positives when the difference between a number and itself is below the threshold, which should not be considered a pair of close elements.\n\nTo fix this issue, you should modify the condition to exclude comparisons between the same numbers by adding an additional check to ensure `i != j` before comparing the two numbers:\n```python\nreturn any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n```\n\nThis adjustment ensures that each number is only compared with distinct numbers in the list, eliminating false positives that would occur when comparing a number with itself. \n\nAfter making this change, the function should correctly check for close elements in the list according to the provided threshold.", prompt_tokens=10, completion_tokens=20, total_tokens=30, prompt_cost=1.5e-05, completion_cost=3.9999999999999996e-05, total_cost=5.4999999999999995e-05, prompt_time=0.5))])
     responses = [
         "```python\ndef has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```",
         "The implementation of the `has_close_elements` function is missing the return statement. The function itself iterates through pairs of numbers in the list and checks if the absolute difference between them falls below the given threshold. However, it lacks a return statement to provide the final result of whether any two numbers are closer than the threshold.\n\nTo fix this issue, the `has_close_elements` function should include a return statement at the end to return the boolean result after checking all pairs of numbers. Here is the revised code with the return statement added:\n\n```python\ndef has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```\n\nBy adding `return` at the beginning of this line, the function will correctly return `True` if any two numbers in the list are closer than the threshold and `False` otherwise.",
@@ -67,87 +68,7 @@ def test_generate() -> None:
         max_interactions=3,
         reset=True,
     )
-
-    assert out == [
-        SelfRefineStepOutput(
-            answer="def has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])",
-            critique="The implementation of the `has_close_elements` function is missing the return statement. The function itself iterates through pairs of numbers in the list and checks if the absolute difference between them falls below the given threshold. However, it lacks a return statement to provide the final result of whether any two numbers are closer than the threshold.\n\nTo fix this issue, the `has_close_elements` function should include a return statement at the end to return the boolean result after checking all pairs of numbers. Here is the revised code with the return statement added:\n\n```python\ndef has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```\n\nBy adding `return` at the beginning of this line, the function will correctly return `True` if any two numbers in the list are closer than the threshold and `False` otherwise.",
-            answer_response=Response(
-                input_text="",
-                output_text="```python\ndef has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```",
-                prompt_tokens=10,
-                completion_tokens=20,
-                total_tokens=30,
-                prompt_cost=1.5e-05,
-                completion_cost=3.9999999999999996e-05,
-                total_cost=5.4999999999999995e-05,
-                prompt_time=0.5,
-            ),
-            critique_response=Response(
-                input_text="",
-                output_text="The implementation of the `has_close_elements` function is missing the return statement. The function itself iterates through pairs of numbers in the list and checks if the absolute difference between them falls below the given threshold. However, it lacks a return statement to provide the final result of whether any two numbers are closer than the threshold.\n\nTo fix this issue, the `has_close_elements` function should include a return statement at the end to return the boolean result after checking all pairs of numbers. Here is the revised code with the return statement added:\n\n```python\ndef has_close_elements(numbers, threshold):\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```\n\nBy adding `return` at the beginning of this line, the function will correctly return `True` if any two numbers in the list are closer than the threshold and `False` otherwise.",
-                prompt_tokens=10,
-                completion_tokens=20,
-                total_tokens=30,
-                prompt_cost=1.5e-05,
-                completion_cost=3.9999999999999996e-05,
-                total_cost=5.4999999999999995e-05,
-                prompt_time=0.5,
-            ),
-        ),
-        SelfRefineStepOutput(
-            answer='from typing import List\n\ndef has_close_elements(numbers: List[float], threshold: float) -> bool:\n    """ Check if in given list of numbers, are any two numbers closer to each other than\n    given threshold.\n    >>> has_close_elements([1.0, 2.0, 3.0], 0.5)\n    False\n    >>> has_close_elements([1.0, 2.8, 3.0, 4.0, 5.0, 2.0], 0.3)\n    True\n    """\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])',
-            critique="The implementation of the `has_close_elements` function has a logical error in the condition used to check if any two numbers in the list are closer to each other than the given threshold.\n\n1. The primary issue lies in the condition used within the `any` function:\n   ```python\n   return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n   ```\n   This condition is meant to check the absolute difference between every pair of numbers (a, b) in the list and return `True` if the difference is less than the threshold for any pair. However, the current implementation fails to consider the case when `a` and `b` are the same number (which implies a difference of 0), potentially leading to incorrect results.\n\n2. In the given test cases:\n   - The case with the numbers `[1.0, 2.0, 3.9, 4.0, 5.0, 2.2]` and threshold `0.3` is expected to return `True` since `2.2` and `3.9` are closer than the threshold.\n   - The case with the same numbers but a threshold of `0.05` is expected to return `False` since the closest numbers are not within the threshold.\n\n3. To address the issue, the condition should be modified to exclude comparing the same number with itself:\n   ```python\n   return any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n   ```\n\nBy adjusting the condition to exclude comparing a number with itself, the function will correctly identify cases where two distinct numbers in the list are closer to each other than the specified threshold, providing accurate outcomes.",
-            answer_response=Response(
-                input_text="",
-                output_text='```python\nfrom typing import List\n\ndef has_close_elements(numbers: List[float], threshold: float) -> bool:\n    """ Check if in given list of numbers, are any two numbers closer to each other than\n    given threshold.\n    >>> has_close_elements([1.0, 2.0, 3.0], 0.5)\n    False\n    >>> has_close_elements([1.0, 2.8, 3.0, 4.0, 5.0, 2.0], 0.3)\n    True\n    """\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n```',
-                prompt_tokens=10,
-                completion_tokens=20,
-                total_tokens=30,
-                prompt_cost=1.5e-05,
-                completion_cost=3.9999999999999996e-05,
-                total_cost=5.4999999999999995e-05,
-                prompt_time=0.5,
-            ),
-            critique_response=Response(
-                input_text="",
-                output_text="The implementation of the `has_close_elements` function has a logical error in the condition used to check if any two numbers in the list are closer to each other than the given threshold.\n\n1. The primary issue lies in the condition used within the `any` function:\n   ```python\n   return any(abs(a - b) < threshold for i, a in enumerate(numbers) for b in numbers[i+1:])\n   ```\n   This condition is meant to check the absolute difference between every pair of numbers (a, b) in the list and return `True` if the difference is less than the threshold for any pair. However, the current implementation fails to consider the case when `a` and `b` are the same number (which implies a difference of 0), potentially leading to incorrect results.\n\n2. In the given test cases:\n   - The case with the numbers `[1.0, 2.0, 3.9, 4.0, 5.0, 2.2]` and threshold `0.3` is expected to return `True` since `2.2` and `3.9` are closer than the threshold.\n   - The case with the same numbers but a threshold of `0.05` is expected to return `False` since the closest numbers are not within the threshold.\n\n3. To address the issue, the condition should be modified to exclude comparing the same number with itself:\n   ```python\n   return any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n   ```\n\nBy adjusting the condition to exclude comparing a number with itself, the function will correctly identify cases where two distinct numbers in the list are closer to each other than the specified threshold, providing accurate outcomes.",
-                prompt_tokens=10,
-                completion_tokens=20,
-                total_tokens=30,
-                prompt_cost=1.5e-05,
-                completion_cost=3.9999999999999996e-05,
-                total_cost=5.4999999999999995e-05,
-                prompt_time=0.5,
-            ),
-        ),
-        SelfRefineStepOutput(
-            answer='from typing import List\n\ndef has_close_elements(numbers: List[float], threshold: float) -> bool:\n    """ Check if in given list of numbers, are any two numbers closer to each other than\n    given threshold.\n    >>> has_close_elements([1.0, 2.0, 3.0], 0.5)\n    False\n    >>> has_close_elements([1.0, 2.8, 3.0, 4.0, 5.0, 2.0], 0.3)\n    True\n    """\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)',
-            critique="The `has_close_elements` function has a logical issue in the way it checks for close elements in the given list of numbers. \n\nHere's the problem: \n\nThe function uses a nested loop with list comprehension to iterate over every pair of numbers in the list and checks if the absolute difference between them is less than the threshold. However, this implementation incorrectly compares every number with every other number in the list, including comparing a number with itself (i.e., when `i == j`). This leads to false positives when the difference between a number and itself is below the threshold, which should not be considered a pair of close elements.\n\nTo fix this issue, you should modify the condition to exclude comparisons between the same numbers by adding an additional check to ensure `i != j` before comparing the two numbers:\n```python\nreturn any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n```\n\nThis adjustment ensures that each number is only compared with distinct numbers in the list, eliminating false positives that would occur when comparing a number with itself. \n\nAfter making this change, the function should correctly check for close elements in the list according to the provided threshold.",
-            answer_response=Response(
-                input_text="",
-                output_text='```python\nfrom typing import List\n\ndef has_close_elements(numbers: List[float], threshold: float) -> bool:\n    """ Check if in given list of numbers, are any two numbers closer to each other than\n    given threshold.\n    >>> has_close_elements([1.0, 2.0, 3.0], 0.5)\n    False\n    >>> has_close_elements([1.0, 2.8, 3.0, 4.0, 5.0, 2.0], 0.3)\n    True\n    """\n    return any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n```',
-                prompt_tokens=10,
-                completion_tokens=20,
-                total_tokens=30,
-                prompt_cost=1.5e-05,
-                completion_cost=3.9999999999999996e-05,
-                total_cost=5.4999999999999995e-05,
-                prompt_time=0.5,
-            ),
-            critique_response=Response(
-                input_text="",
-                output_text="The `has_close_elements` function has a logical issue in the way it checks for close elements in the given list of numbers. \n\nHere's the problem: \n\nThe function uses a nested loop with list comprehension to iterate over every pair of numbers in the list and checks if the absolute difference between them is less than the threshold. However, this implementation incorrectly compares every number with every other number in the list, including comparing a number with itself (i.e., when `i == j`). This leads to false positives when the difference between a number and itself is below the threshold, which should not be considered a pair of close elements.\n\nTo fix this issue, you should modify the condition to exclude comparisons between the same numbers by adding an additional check to ensure `i != j` before comparing the two numbers:\n```python\nreturn any(abs(a - b) < threshold for i, a in enumerate(numbers) for j, b in enumerate(numbers) if i != j)\n```\n\nThis adjustment ensures that each number is only compared with distinct numbers in the list, eliminating false positives that would occur when comparing a number with itself. \n\nAfter making this change, the function should correctly check for close elements in the list according to the provided threshold.",
-                prompt_tokens=10,
-                completion_tokens=20,
-                total_tokens=30,
-                prompt_cost=1.5e-05,
-                completion_cost=3.9999999999999996e-05,
-                total_cost=5.4999999999999995e-05,
-                prompt_time=0.5,
-            ),
-        ),
-    ]
+    assert out == gt_out
 
 
 def test_generate_answer() -> None:
